+++ python/paddle/nn/functional/__init__.py
@@ -94,7 +94,7 @@ from .flash_attention import (
     scaled_dot_product_attention,
     sdp_kernel,  # noqa: F401
 )
-from .fused_transpose_split_quant import fused_transpose_split_quant
+from .fp8 import fused_transpose_split_quant
 from .input import (
     embedding,
     embedding_renorm_,  # noqa: F401
